+++ src/test/kotlin/dev/tailwindrainbow/intellij/application/settings/SettingsFormMapperTest.kt
@@ -205,5 +205,6 @@ class SettingsFormMapperTest {
         templateTags = "tw",
         ignoredPrefixModifiers = "group",
         supportedExtensions = supportedExtensions,
+        readsClassLikeStrings = true,
     )
 }